Add RendererCreationEvent
