+++ patches/minecraft/net/minecraft/client/renderer/entity/LivingRenderer.java.patch
@@ -1,14 +1,18 @@
 --- a/net/minecraft/client/renderer/entity/LivingRenderer.java
 +++ b/net/minecraft/client/renderer/entity/LivingRenderer.java
-@@ -40,6 +40,7 @@
+@@ -37,9 +37,11 @@
+       super(p_i50965_1_);
+       this.field_77045_g = p_i50965_2_;
+       this.field_76989_e = p_i50965_3_;
++      net.minecraftforge.fml.ModLoader.get().postEvent(new net.minecraftforge.client.event.LivingRendererEvent.RendererCreationEvent<>(this));
     }
  
     public final boolean func_177094_a(LayerRenderer<T, M> p_177094_1_) {
-+      if(net.minecraftforge.fml.ModLoader.get().postEvent(new net.minecraftforge.client.event.LayerRendererEvent.Add<>(this, p_177094_1_))) return false;
++      if(net.minecraftforge.fml.ModLoader.get().postEvent(new net.minecraftforge.client.event.LivingRendererEvent.LayerEvent.Add<>(this, p_177094_1_))) return false;
        return this.field_177097_h.add(p_177094_1_);
     }
  
-@@ -48,14 +49,17 @@
+@@ -48,14 +50,17 @@
     }
  
     public void func_225623_a_(T p_225623_1_, float p_225623_2_, float p_225623_3_, MatrixStack p_225623_4_, IRenderTypeBuffer p_225623_5_, int p_225623_6_) {
@@ -28,7 +32,7 @@
           LivingEntity livingentity = (LivingEntity)p_225623_1_.func_184187_bx();
           f = MathHelper.func_219805_h(p_225623_3_, livingentity.field_70760_ar, livingentity.field_70761_aq);
           f2 = f1 - f;
-@@ -92,7 +96,7 @@
+@@ -92,7 +97,7 @@
        p_225623_4_.func_227861_a_(0.0D, (double)-1.501F, 0.0D);
        float f8 = 0.0F;
        float f5 = 0.0F;
@@ -37,7 +41,7 @@
           f8 = MathHelper.func_219799_g(p_225623_3_, p_225623_1_.field_184618_aE, p_225623_1_.field_70721_aZ);
           f5 = p_225623_1_.field_184619_aG - p_225623_1_.field_70721_aZ * (1.0F - p_225623_3_);
           if (p_225623_1_.func_70631_g_()) {
-@@ -125,6 +129,7 @@
+@@ -125,6 +130,7 @@
  
        p_225623_4_.func_227865_b_();
        super.func_225623_a_(p_225623_1_, p_225623_2_, p_225623_3_, p_225623_4_, p_225623_5_, p_225623_6_);
@@ -45,14 +49,14 @@
     }
  
     @Nullable
-@@ -254,4 +259,10 @@
+@@ -254,4 +260,10 @@
           return Minecraft.func_71382_s() && p_177070_1_ != minecraft.func_175606_aa() && flag && !p_177070_1_.func_184207_aI();
        }
     }
 +   /* ============= FORGE =============*/
 +   public List<LayerRenderer<T, M>> getLayers() {return Lists.newArrayList(field_177097_h);}
 +   public boolean removeLayer(LayerRenderer<T, M> layer) {
-+      if(net.minecraftforge.fml.ModLoader.get().postEvent(new net.minecraftforge.client.event.LayerRendererEvent.Remove<>(this, layer))) return false;
++      if(net.minecraftforge.fml.ModLoader.get().postEvent(new net.minecraftforge.client.event.LivingRendererEvent.LayerEvent.Remove<>(this, layer))) return false;
 +      return field_177097_h.remove(layer);}
 +   
  }
